move towards the new fastboot image
cecemel/ember-fastboot-proxy-service is no longer maintained,
redpencil/fastboot-app-server is now preferred.

the new image works a bit differently because it assumes the frontend is behind
the dispatcher. as such it only provides a reverse proxy for fastboot but not
for any direct requests.

this is in turn means we need an application adapter to discover the fastboot
case and reroute our requests to the appriopate backend