Reviewed RDFa content on templates and models
Some models needed extra RDFa properties and the RDFa on the templates
has been reviewed. We use rdfa-helpers as much as possible to eliminate
mistakes and to provide better flexibility.
This does not include a complete check on the quality and correctnes of
the RDFa data, but some minor mistakes have been fixed.